feat: 모달을 관리하는 useModal 커스텀 훅 생성
