+++ src/hooks/useModal.ts
@@ -1,0 +1,13 @@
+import { useState } from 'react';
+
+export const useModal = () => {
+  const [isModalOpen, setIsModalOpen] = useState(false);
+  const openModal = () => {
+    setIsModalOpen(true);
+  };
+
+  const closeModal = () => {
+    setIsModalOpen(false);
+  };
+  return { isModalOpen, openModal, closeModal };
+};
